Change return type for RSSI from float to int for consistency
Test: Successfully built.
Bug: 189954300

+++ nearby/framework/java/android/nearby/FastPairDevice.java
@@ -92,7 +92,7 @@ public class FastPairDevice extends NearbyDevice implements Parcelable {
     }
 
     @Override
-    public float getRssi() {
+    public int getRssi() {
         return mRssi;
     }
